requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
httpx==0.27.2
trafilatura==1.12.2
lxml_html_clean==0.4.3
